Fixing some method call to an object that is geting from GLOBALS["plugins"] array. The problem with that is that: in each cicle you will never know if the method exists in the current object, that's why I consider necessary to "ask" if the methos exists, and then, call it.

+++ commonlib/lib/userlib.php
@@ -60,8 +60,13 @@ function deleteUser($id) {
   Sql_Query(sprintf('delete from %s where id = %d',$tables["user"],$id));
   Sql_Query(sprintf('delete from %s where userid = %d',$tables["user_history"],$id));
   ### allow plugins to delete their data
+
   foreach ($GLOBALS['plugins'] as $plugin) {
-    $plugin->deleteUser($id);
+    //$plugin->deleteUser($id);
+
+    if (method_exists($plugin, "deleteUser")) {
+       $plugin->deleteUser($id);
+    }
   }
 }
